One-to-Many: Nested Resources
Nested the review resource inside of a movie resource so that the router
recognizes requests for /movies/1/reviews, for example.

Defined an index action in the ReviewsController that fetches all the
reviews for the movie specified in the URL.

Created a corresponding index.html.erb template that generates an HTML
list of the movie's reviews.

Generated a "Reviews" link on the movie show page so it's easy to
navigate to the list of reviews for any movie.