One-to-Many: Forms
Generated a "Write Review" link on the movie show page so it's easy to
get to the review form for any movie.

Defined a new action in the ReviewsController that finds the movie being
reviewed and uses it to instantiate a new Review object to bind to the
form.

Created a corresponding new.html.erb template that generates an HTML
form with fields for the review.

Defined a create action in the ReviewsController that uses the form data
to create a new review in the database and associates it with the
respective movie.